+++ gradle-plugins/cordformation/src/main/kotlin/net/corda/plugins/Cordform.kt
@@ -20,7 +20,8 @@ import java.util.jar.JarInputStream
  * See documentation for examples.
  */
 @Suppress("unused")
-open class Cordform : DefaultTask() {
+open class Cordform @Inject constructor(private val objectFactory: ObjectFactory) : DefaultTask() {
+
     private companion object {
         val nodeJarName = "corda.jar"
         private val defaultDirectory: Path = Paths.get("build", "nodes")
@@ -50,7 +51,7 @@ open class Cordform : DefaultTask() {
      */
     @Suppress("MemberVisibilityCanPrivate")
     fun node(configureClosure: Closure<in Node>) {
-        nodes += project.configure(Node(project), configureClosure) as Node
+        nodes += project.configure(objectFactory.newInstance(Node::class.java, project), configureClosure) as Node
     }
 
     /**
@@ -60,10 +61,13 @@ open class Cordform : DefaultTask() {
      */
     @Suppress("MemberVisibilityCanPrivate")
     fun node(configureFunc: Node.() -> Any?): Node {
-        val node = Node(project).apply { configureFunc() }
+        val node = objectFactory.newInstance(Node::class.java, project).apply {
+            configureFunc()
+        }
         nodes += node
         return node
     }
+
 
     /**
      * Returns a node by name.
@@ -79,7 +83,7 @@ open class Cordform : DefaultTask() {
     private fun installRunScript() {
         project.copy {
             it.apply {
-                from(Cordformation.getPluginFile(project, "net/corda/plugins/runnodes.jar"))
+                from(Cordformation.getPluginFile("runnodes.jar"))
                 fileMode = Cordformation.executableFileMode
                 into("$directory/")
             }
@@ -87,7 +91,7 @@ open class Cordform : DefaultTask() {
 
         project.copy {
             it.apply {
-                from(Cordformation.getPluginFile(project, "net/corda/plugins/runnodes"))
+                from(Cordformation.getPluginFile("runnodes"))
                 // Replaces end of line with lf to avoid issues with the bash interpreter and Windows style line endings.
                 filter(mapOf("eol" to FixCrLfFilter.CrLf.newInstance("lf")), FixCrLfFilter::class.java)
                 fileMode = Cordformation.executableFileMode
@@ -97,7 +101,7 @@ open class Cordform : DefaultTask() {
 
         project.copy {
             it.apply {
-                from(Cordformation.getPluginFile(project, "net/corda/plugins/runnodes.bat"))
+                from(Cordformation.getPluginFile("runnodes.bat"))
                 into("$directory/")
             }
         }