+++ gradle-plugins/cordformation/src/main/kotlin/net/corda/plugins/Node.kt
@@ -3,8 +3,17 @@ package net.corda.plugins
 import com.typesafe.config.ConfigFactory
 import com.typesafe.config.ConfigRenderOptions
 import com.typesafe.config.ConfigValueFactory
+import groovy.lang.Closure
+import com.typesafe.config.ConfigFactory
+import com.typesafe.config.ConfigRenderOptions
+import com.typesafe.config.ConfigValueFactory
 import net.corda.cordform.CordformNode
+import org.apache.commons.io.FilenameUtils
+import org.bouncycastle.asn1.x500.X500Name
+import org.bouncycastle.asn1.x500.style.BCStyle
+import org.gradle.api.GradleException
 import org.gradle.api.Project
+import org.gradle.api.model.ObjectFactory
 import java.io.File
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
@@ -13,7 +22,9 @@ import java.nio.file.Path
 /**
  * Represents a node that will be installed.
  */
-class Node(private val project: Project) : CordformNode() {
+open class Node @Inject constructor(private val project: Project) : CordformNode() {
+    private data class ResolvedCordapp(val jarFile: File, val config: String?)
+
     companion object {
         @JvmStatic
         val webJarName = "corda-webserver.jar"
@@ -27,8 +38,17 @@ class Node(private val project: Project) : CordformNode() {
      * @note Your app will be installed by default and does not need to be included here.
      * @note Type is any due to gradle's use of "GStrings" - each value will have "toString" called on it
      */
-    var cordapps = mutableListOf<Any>()
-    internal var additionalCordapps = mutableListOf<File>()
+    var cordapps: MutableList<Any>
+        get() = internalCordapps as MutableList<Any>
+        @Deprecated("Use cordapp instead - setter will be removed by Corda V4.0")
+        set(value) {
+            value.forEach {
+                cordapp(it.toString())
+            }
+        }
+
+    private val internalCordapps = mutableListOf<Cordapp>()
+    private val releaseVersion = project.rootProject.ext<String>("corda_release_version")
     internal lateinit var nodeDir: File
         private set
     internal lateinit var rootDir: File
@@ -59,17 +79,94 @@ class Node(private val project: Project) : CordformNode() {
      *
      * @param sshdPort The port for SSH server to listen on
      */
-    fun sshdPort(sshdPort: Int?) {
-        config = config.withValue("sshd.port", ConfigValueFactory.fromAnyRef(sshdPort))
+    fun sshdPort(sshdPort: Int) {
+        config = config.withValue("sshdAddress",
+                ConfigValueFactory.fromAnyRef("$DEFAULT_HOST:$sshdPort"))
     }
+
+    /**
+     * Configures the default cordapp automatically added to this node
+     *
+     * @param configureClosure A groovy closure to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(configureClosure: Closure<in Cordapp>): Cordapp {
+        project.configure(builtCordapp, configureClosure) as Cordapp
+        return builtCordapp
+    }
+
+    /**
+     * Install a cordapp to this node
+     *
+     * @param coordinates The coordinates of the [Cordapp]
+     * @param configureClosure A groovy closure to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(coordinates: String, configureClosure: Closure<in Cordapp>): Cordapp {
+        val cordapp = project.configure(Cordapp(coordinates), configureClosure) as Cordapp
+        internalCordapps += cordapp
+        return cordapp
+    }
+
+    /**
+     * Install a cordapp to this node
+     *
+     * @param cordappProject A project that produces a cordapp JAR
+     * @param configureClosure A groovy closure to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(cordappProject: Project, configureClosure: Closure<in Cordapp>): Cordapp {
+        val cordapp = project.configure(Cordapp(cordappProject), configureClosure) as Cordapp
+        internalCordapps += cordapp
+        return cordapp
+    }
+
+    /**
+     * Install a cordapp to this node
+     *
+     * @param cordappProject A project that produces a cordapp JAR
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(cordappProject: Project): Cordapp {
+        return Cordapp(cordappProject).apply {
+            internalCordapps += this
+        }
+    }
+
+    /**
+     * Install a cordapp to this node
+     *
+     * @param coordinates The coordinates of the [Cordapp]
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(coordinates: String): Cordapp {
+        return Cordapp(coordinates).apply {
+            internalCordapps += this
+        }
+    }
+
+    /**
+     * Install a cordapp to this node
+     *
+     * @param configureFunc A lambda to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(coordinates: String, configureFunc: Cordapp.() -> Unit): Cordapp {
+        return Cordapp(coordinates).apply {
+            configureFunc()
+            internalCordapps += this
+        }
+    }
+
 
     internal fun build() {
         if (config.hasPath("webAddress")) {
             installWebserverJar()
         }
-        installAgentJar()
         installBuiltCordapp()
         installCordapps()
+        installConfig()
+        appendOptionalConfig()
     }
 
     internal fun rootDir(rootDir: Path) {
@@ -111,16 +208,19 @@ class Node(private val project: Project) : CordformNode() {
     }
 
     /**
-     * Installs this project's cordapp to this directory.
+     * Installs the jolokia monitoring agent JAR to the node/drivers directory
      */
-    private fun installBuiltCordapp() {
+    private fun installCordapps() {
+        val cordapps = getCordappList()
         val cordappsDir = File(nodeDir, "cordapps")
         project.copy {
             it.apply {
-                from(project.tasks.getByName("jar"))
-                into(cordappsDir)
+                from(cordapps.map { it.jarFile })
+                into(project.file(cordappsDir))
             }
         }
+
+        installCordappConfigs(cordapps)
     }
 
     /**
@@ -146,6 +246,14 @@ class Node(private val project: Project) : CordformNode() {
         }
     }
 
+    private fun installCordappConfigs(cordapps: Collection<ResolvedCordapp>) {
+        val cordappsDir = project.file(File(nodeDir, "cordapps"))
+        cordappsDir.mkdirs()
+        cordapps.filter { it.config != null }
+                .map { Pair<String, String>("${FilenameUtils.removeExtension(it.jarFile.name)}.conf", it.config!!) }
+                .forEach { project.file(File(cordappsDir, it.first)).writeText(it.second) }
+    }
+
     private fun createTempConfigFile(): File {
         val options = ConfigRenderOptions
                 .defaults()
@@ -166,14 +274,25 @@ class Node(private val project: Project) : CordformNode() {
     /**
      * Installs the configuration file to the root directory and detokenises it.
      */
-    internal fun installConfig() {
-        configureProperties()
-        val tmpConfFile = createTempConfigFile()
-        appendOptionalConfig(tmpConfFile)
+    private fun installConfig() {
+        val options = ConfigRenderOptions
+                .defaults()
+                .setOriginComments(false)
+                .setComments(false)
+                .setFormatted(true)
+                .setJson(false)
+        val configFileText = config.root().render(options).split("\n").toList()
+
+        // Need to write a temporary file first to use the project.copy, which resolves directories correctly.
+        val tmpDir = File(project.buildDir, "tmp")
+        tmpDir.mkdir()
+        val tmpConfFile = File(tmpDir, "node.conf")
+        Files.write(tmpConfFile.toPath(), configFileText, StandardCharsets.UTF_8)
+
         project.copy {
             it.apply {
                 from(tmpConfFile)
-                into(rootDir)
+                into(nodeDir)
             }
         }
     }
@@ -181,7 +300,7 @@ class Node(private val project: Project) : CordformNode() {
     /**
      * Appends installed config file with properties from an optional file.
      */
-    private fun appendOptionalConfig(confFile: File) {
+    private fun appendOptionalConfig() {
         val optionalConfig: File? = when {
             project.findProperty(configFileProperty) != null -> //provided by -PconfigFile command line property when running Gradle task
                 File(project.findProperty(configFileProperty) as String)
@@ -193,6 +312,7 @@ class Node(private val project: Project) : CordformNode() {
             if (!optionalConfig.exists()) {
                 project.logger.error("$configFileProperty '$optionalConfig' not found")
             } else {
+                val confFile = File(project.buildDir.path + "/../" + nodeDir, "node.conf")
                 confFile.appendBytes(optionalConfig.readBytes())
             }
         }
@@ -213,16 +333,63 @@ class Node(private val project: Project) : CordformNode() {
     }
 
     /**
+     * Add a cordapp to this node
+     *
+     * @param configureClosure A groovy closure to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(configureClosure: Closure<in Cordapp>): Cordapp {
+        val cordapp = project.configure(objectFactory.newInstance(Cordapp::class.java), configureClosure) as Cordapp
+        addCordapp(cordapp)
+        return cordapp
+    }
+
+    /**
+     * Add a cordapp to this node
+     *
+     * @param configureFunc A lambda to configure a [Cordapp] object
+     * @return The created and inserted [Cordapp]
+     */
+    fun cordapp(configureFunc: Cordapp.() -> Unit): Cordapp {
+        val cordapp = objectFactory.newInstance(Cordapp::class.java).apply { configureFunc() }
+        addCordapp(cordapp)
+        return cordapp
+    }
+
+
+    /**
      * Gets a list of cordapps based on what dependent cordapps were specified.
      *
      * @return List of this node's cordapps.
      */
-    private fun getCordappList(): Collection<File> {
-        // Cordapps can sometimes contain a GString instance which fails the equality test with the Java string
-        @Suppress("RemoveRedundantCallsOfConversionMethods")
-        val cordapps: List<String> = cordapps.map { it.toString() }
-        return project.configuration("cordapp").files {
-            cordapps.contains(it.group + ":" + it.name + ":" + it.version)
+    private fun getCordappList(): Collection<ResolvedCordapp> =
+            internalCordapps.map { cordapp -> resolveCordapp(cordapp) } + resolveBuiltCordapp()
+
+    private fun resolveCordapp(cordapp: Cordapp): ResolvedCordapp {
+        val cordappConfiguration = project.configuration("cordapp")
+        val cordappName = if(cordapp.project != null) cordapp.project.name else cordapp.coordinates
+        val cordappFile = cordappConfiguration.files {
+            when {
+                it is ProjectDependency -> it.dependencyProject == cordapp.project!!
+                cordapp.coordinates != null -> {
+                    // Cordapps can sometimes contain a GString instance which fails the equality test with the Java string
+                    @Suppress("RemoveRedundantCallsOfConversionMethods")
+                    val coordinates = cordapp.coordinates.toString()
+                    coordinates == (it.group + ":" + it.name + ":" + it.version)
+                }
+                else -> false
+            }
         }
+
+        return when {
+            cordappFile.size == 0 -> throw GradleException("Cordapp $cordappName not found in cordapps configuration.")
+            cordappFile.size > 1 -> throw GradleException("Multiple files found for $cordappName")
+            else -> ResolvedCordapp(cordappFile.single(), cordapp.config)
+        }
+    }
+
+    private fun resolveBuiltCordapp(): ResolvedCordapp {
+        val projectCordappFile = project.tasks.getByName("jar").outputs.files.singleFile
+        return ResolvedCordapp(projectCordappFile, builtCordapp.config)
     }
 }